Crate osdk_frame_allocator

An implementation of the global physical memory frame allocator for OSTD based kernels.

Background

OSTD has provided a page allocator interface, namely GlobalFrameAllocator and global_frame_allocator procedure macro, allowing users to plug in their own frame allocator into the kernel safely. You can refer to the ostd::mm::frame::allocator module for detailed introduction.

Introduction

This crate is an implementation of a scalable and efficient global frame allocator based on the buddy system. It is by default shipped with OSDK for users that don’t have special requirements on the frame allocator.

Macros

fast_smp_counter

Defines a static fast SMP counter.

Structs

FrameAllocator

The global frame allocator provided by OSDK.

Statics

TOTAL_FREE_SIZE

The total size of free memory.

Functions

load_total_free_size

Loads the total size (in bytes) of free memory in the allocator.